+++ packages/react-dashboard/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.9.0 (2026-07-14)
+
+### Features
+
+- **react-dashboard:** add BigNumberSparkline card for lineChart type ([#1124](https://github.com/ttoss/ttoss/issues/1124)) ([3917371](https://github.com/ttoss/ttoss/commit/3917371afeb9e8bde26cabe13d7284119948786d))
+
 ## [0.8.4](https://github.com/ttoss/ttoss/compare/@ttoss/react-dashboard@0.8.3...@ttoss/react-dashboard@0.8.4) (2026-06-28)
 
 **Note:** Version bump only for package @ttoss/react-dashboard